fix(001): security review hardening — close all 16 findings (5 P1, 6 P2, 5 P3)
P1: claim_passport serialized per phone (advisory xact lock) + merged_into
chain flattening; unique_violation → INVALID_STATE (never raw 23505); OTP
brute-force lockout (5 attempts, autonomous counter) + 5/hr send cap;
STRIPE_WEBHOOK_SECRET fails closed; all §5 views security_invoker.

P2: redeem_perk advisory lock kills the TOCTOU double-redeem; webhook dedup
is claim-first INSERT with release-on-failure; OTP consume is a single
guarded UPDATE; merge dedupes colliding impressions/unlocks instead of
orphaning them (redemption repoint kept unconditional by design — audit
history, correct D-018 reset anchor); gate_metrics_evaluated revoked from
app roles.

P3: business-detail perk progress uses since-last-redemption rule; regulars
sort numerically; perk_progress_count revoked from app roles; staff_check_in
phone race caught; regional_milestones natural key makes seed idempotent.

Tests: +12 regression tests (review-hardening.test.ts) covering all six
review gaps. 124/124 vitest, 14/14 Playwright, tsc + brand clean. D-021.

+++ supabase/migrations/0013_rpc_regulars.sql
@@ -32,7 +32,8 @@ begin
     raise exception 'FORBIDDEN' using errcode = 'P0001';
   end if;
 
-  select coalesce(jsonb_agg(row order by row->>'visits' desc), '[]'::jsonb) into v_out
+  -- ::int — '->>'' yields text and "9" sorts above "10" lexicographically (review P3-13).
+  select coalesce(jsonb_agg(row order by (row->>'visits')::int desc), '[]'::jsonb) into v_out
   from (
     select jsonb_build_object(
       'patron_ref', s.patron_id,
